Remove the auxiliary RulesetConfigs tuple struct and move the deserialization stuff to another file.
