+++ crates/bins/src/bin/datadog-static-analyzer.rs
@@ -236,7 +236,7 @@ fn main() -> Result<()> {
             exit(1);
         }
 
-        let rulesets = conf.rulesets.0.keys().cloned().collect_vec();
+        let rulesets = conf.rulesets.keys().cloned().collect_vec();
         let rules_from_api = get_rules_from_rulesets(&rulesets, use_staging);
         rules.extend(rules_from_api.context("error when reading rules from API")?);
